+++ cmd/neofs-node/config/internal/validate/config.go
@@ -84,14 +84,6 @@ type valideConfig struct {
 		Proxy      string `mapstructure:"proxy"`
 	} `mapstructure:"contracts"`
 
-	Morph struct {
-		DialTimeout         time.Duration `mapstructure:"dial_timeout"`
-		CacheTTL            time.Duration `mapstructure:"cache_ttl"`
-		ReconnectionsNumber int           `mapstructure:"reconnections_number"`
-		ReconnectionsDelay  time.Duration `mapstructure:"reconnections_delay"`
-		Endpoints           []string      `mapstructure:"endpoints"`
-	} `mapstructure:"morph"`
-
 	FSChain struct {
 		DialTimeout         time.Duration `mapstructure:"dial_timeout"`
 		CacheTTL            time.Duration `mapstructure:"cache_ttl"`